<fix>[storage]: fix mark vol as snapshot rollback
handle NPE

Resolves: ZSTAC-63723

+++ storage/src/main/java/org/zstack/storage/snapshot/VolumeSnapshotManagerImpl.java
@@ -671,6 +671,9 @@ public class VolumeSnapshotManagerImpl extends AbstractService implements
     @Transactional
     private void rollbackSnapshot(String uuid) {
         VolumeSnapshotVO vo = dbf.getEntityManager().find(VolumeSnapshotVO.class, uuid);
+        if (vo == null) {
+            return;
+        }
 
         dbf.getEntityManager().remove(vo);
